// packages/declare/src/generated/user.ts
export const UserBase_2_40 = z.object({
  accountExpiry: z.string().optional(),
  birthday: z.string().optional(),
  code: z.string().max(50).optional(),
  dataViewMaxOrganisationUnitLevel: z.number().int().optional(),
  dataViewOrganisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  disabled: z.boolean().optional(),
  education: z.string().optional(),
  email: z.string().email().optional(),
  employer: z.string().max(160).optional(),
  externalAuth: z.boolean().optional(),
  facebookMessenger: z.string().max(255).optional(),
  firstName: z.string().max(160).optional(),
  gender: z.string().max(50).optional(),
  interests: z.string().optional(),
  introduction: z.string().optional(),
  invitation: z.boolean().optional(),
  jobTitle: z.string().max(160).optional(),
  languages: z.string().optional(),
  lastCheckedInterpretations: z.string().optional(),
  lastLogin: z.string().optional(),
  ldapId: z.string().optional(),
  nationality: z.string().max(160).optional(),
  openId: z.string().optional(),
  organisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  password: z.string().optional(),
  passwordLastUpdated: z.string().optional(),
  phoneNumber: z.string().max(80).optional(),
  selfRegistered: z.boolean().optional(),
  skype: z.string().max(255).optional(),
  surname: z.string().max(160).optional(),
  teiSearchOrganisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  telegram: z.string().max(255).optional(),
  twitter: z.string().max(255).optional(),
  username: z.string().max(255).optional(),
  userRoles: z.array(refSchema('UserRole')).optional(),
  welcomeMessage: z.string().optional(),
  whatsApp: z.string().max(255).optional(),
})

export const UserBase_2_41 = z.object({
  accountExpiry: z.string().optional(),
  birthday: z.string().optional(),
  code: z.string().max(50).optional(),
  dataViewMaxOrganisationUnitLevel: z.number().int().optional(),
  dataViewOrganisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  disabled: z.boolean().optional(),
  education: z.string().optional(),
  email: z.string().email().optional(),
  employer: z.string().max(160).optional(),
  externalAuth: z.boolean().optional(),
  facebookMessenger: z.string().max(255).optional(),
  firstName: z.string().max(160).optional(),
  gender: z.string().max(50).optional(),
  interests: z.string().optional(),
  introduction: z.string().optional(),
  invitation: z.boolean().optional(),
  jobTitle: z.string().max(160).optional(),
  languages: z.string().optional(),
  lastCheckedInterpretations: z.string().optional(),
  lastLogin: z.string().optional(),
  ldapId: z.string().optional(),
  nationality: z.string().max(160).optional(),
  openId: z.string().optional(),
  organisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  password: z.string().optional(),
  passwordLastUpdated: z.string().optional(),
  phoneNumber: z.string().max(80).optional(),
  selfRegistered: z.boolean().optional(),
  skype: z.string().max(255).optional(),
  surname: z.string().max(160).optional(),
  teiSearchOrganisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  telegram: z.string().max(255).optional(),
  twitter: z.string().max(255).optional(),
  username: z.string().max(255).optional(),
  userRoles: z.array(refSchema('UserRole')).optional(),
  welcomeMessage: z.string().optional(),
  whatsApp: z.string().max(255).optional(),
})

export const UserBase_2_42 = z.object({
  accountExpiry: z.string().optional(),
  birthday: z.string().optional(),
  code: z.string().max(50).optional(),
  dataViewMaxOrganisationUnitLevel: z.number().int().optional(),
  dataViewOrganisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  disabled: z.boolean().optional(),
  education: z.string().optional(),
  email: z.string().email().optional(),
  emailVerificationToken: z.string().max(255).optional(),
  employer: z.string().max(160).optional(),
  externalAuth: z.boolean().optional(),
  facebookMessenger: z.string().max(255).optional(),
  firstName: z.string().max(160),
  gender: z.string().max(50).optional(),
  interests: z.string().optional(),
  introduction: z.string().optional(),
  invitation: z.boolean().optional(),
  jobTitle: z.string().max(160).optional(),
  languages: z.string().optional(),
  lastCheckedInterpretations: z.string().optional(),
  lastLogin: z.string().optional(),
  ldapId: z.string().optional(),
  nationality: z.string().max(160).optional(),
  openId: z.string().optional(),
  organisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  password: z.string().optional(),
  passwordLastUpdated: z.string().optional(),
  phoneNumber: z.string().max(80).optional(),
  selfRegistered: z.boolean().optional(),
  skype: z.string().max(255).optional(),
  surname: z.string().max(160),
  teiSearchOrganisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  telegram: z.string().max(255).optional(),
  twitter: z.string().max(255).optional(),
  username: z.string().max(255).optional(),
  userRoles: z.array(refSchema('UserRole')).optional(),
  verifiedEmail: z.string().max(255).optional(),
  welcomeMessage: z.string().optional(),
  whatsApp: z.string().max(255).optional(),
})

export const UserBaseByTarget = {
  '2.40': UserBase_2_40,
  '2.41': UserBase_2_41,
  '2.42': UserBase_2_42,
} as const